Module spin_sdk::http

Helpers for building Spin wasi-http components.

Modules

conversions

Traits for converting between the various types

responses

Helper functions for creating responses

Macros

http_router

A macro to help with constructing a Router from a stream of tokens.

Structs

Body

A Body extractor

Fields

This following block defines the fields resource which corresponds to HTTP standard Fields.

HeaderValue

A header value.

IncomingRequest

The following block defines the incoming-request and outgoing-request resource types that correspond to HTTP standard Requests.

IncomingResponse

The following block defines the incoming-response and outgoing-response resource types that correspond to HTTP standard Responses.

Json

A Json extractor

JsonBodyError

An error parsing a JSON body

NonUtf8BodyError

An error when the body is not UTF-8

OutgoingBody

OutgoingRequest

OutgoingResponse

Request

A unified request object that can represent both incoming and outgoing requests.

RequestBuilder

A request builder

Response

A unified response object that can represent both outgoing and incoming responses.

ResponseBuilder

A builder for `Response``

ResponseOutparam

The out param for setting an OutgoingResponse

Router

Exports HTTP Router items. The Spin SDK HTTP router.

Enums

Error

This type enumerates the different kinds of errors that may occur when initially returning a response.

Method

This type corresponds to HTTP standard Methods.

Scheme

This type corresponds to HTTP standard Related Schemes.

SendError

An error encountered when performing an HTTP request

Traits

IntoResponse

A trait for any type that can be turned into a Response

Functions

send

Send an outgoing request

Type Aliases

Headers

Params

Exports HTTP Router items. Route parameters extracted from a URI that match a route pattern.

StatusCode

Trailers