Module spin_sdk::http::conversions

Traits for converting between the various types

Traits

FromBody

A trait from converting from a body

IntoBody

A trait for any type that can be turned into a Response body

IntoResponse

A trait for any type that can be turned into a Response

IntoStatusCode

A trait for any type that can be turned into a Response status code

TryFromBody

A trait for converting from a body or failing

TryFromRequest

A trait for any type that can be constructor from a Request

TryIntoBody

A trait for any type that can be turned into a Response body or fail

TryNonRequestFromRequest

A hack that allows us to do blanket impls for T where T: TryFromRequest for all types T except for Request.