Struct spin_sdk::http::Request

pub struct Request { /* private fields */ }

A unified request object that can represent both incoming and outgoing requests.

This should be used in favor of IncomingRequest and OutgoingRequest when there is no need for streaming bodies.

Implementations

impl Request

pub fn new(method: Method, uri: impl Into<String>) -> Self

Creates a new request from a method and uri

pub fn builder() -> RequestBuilder

Creates a RequestBuilder

pub fn get(uri: impl Into<String>) -> RequestBuilder

Creates a RequestBuilder to GET the given uri

pub fn post(uri: impl Into<String>, body: impl IntoBody) -> RequestBuilder

Creates a RequestBuilder to POST the given body to uri

pub fn method(&self) -> &Method

The request method

pub fn uri(&self) -> &str

The request uri

pub fn path(&self) -> &str

The request uri path

pub fn query(&self) -> &str

The request uri query

pub fn headers(&self) -> impl Iterator<Item = (&str, &HeaderValue)>

The request headers

pub fn header(&self, name: &str) -> Option<&HeaderValue>

Return a header value

Will return None if the header does not exist.

pub fn body(&self) -> &[u8]

The request body

pub fn body_mut(&mut self) -> &mut Vec<u8>

The request body

pub fn into_body(self) -> Vec<u8>

Consume this type and return its body

pub fn path_and_query(&self) -> Option<&str>

The request path and query combined