Struct spin_sdk::http::Response

pub struct Response { /* private fields */ }

A unified response object that can represent both outgoing and incoming responses.

This should be used in favor of OutgoingResponse and IncomingResponse when there is no need for streaming bodies.

Implementations

impl Response

pub fn new(status: impl IntoStatusCode, body: impl IntoBody) -> Self

Create a new response from a status and body

pub fn status(&self) -> &StatusCode

The response status

pub fn headers(&self) -> impl Iterator<Item = (&str, &HeaderValue)>

The request headers

pub fn header(&self, name: &str) -> Option<&HeaderValue>

Return a header value

Will return None if the header does not exist.

pub fn body(&self) -> &[u8]

The response body

pub fn body_mut(&mut self) -> &mut Vec<u8>

The response body

pub fn into_body(self) -> Vec<u8>

Consume this type and return its body

pub fn into_builder(self) -> ResponseBuilder

Converts this response into a ResponseBuilder. This can be used to update a response before passing it on.

pub fn builder() -> ResponseBuilder

Creates a ResponseBuilder