Struct spin_sdk::http::Router

pub struct Router { /* private fields */ }

Exports HTTP Router items. The Spin SDK HTTP router.

Implementations

impl Router

pub fn handle<R>(&self, request: R) -> Response
where R: TryIntoRequest, R::Error: IntoResponse,

Synchronously dispatches a request to the appropriate handler along with the URI parameters.

pub async fn handle_async<R>(&self, request: R) -> Response
where R: TryIntoRequest, R::Error: IntoResponse,

Asynchronously dispatches a request to the appropriate handler along with the URI parameters.

pub fn any<F, Req, Resp>(&mut self, path: &str, handler: F)
where F: Fn(Req, Params) -> Resp + 'static, Req: TryFromRequest + 'static, Req::Error: IntoResponse + 'static, Resp: IntoResponse + 'static,

Register a handler at the path for all methods.

pub fn any_async<F, Fut, I, O>(&mut self, path: &str, handler: F)
where F: Fn(I, Params) -> Fut + 'static, Fut: Future<Output = O> + 'static, I: TryFromRequest + 'static, I::Error: IntoResponse + 'static, O: IntoResponse + 'static,

Register an async handler at the path for all methods.

pub fn add<F, Req, Resp>(&mut self, path: &str, method: Method, handler: F)
where F: Fn(Req, Params) -> Resp + 'static, Req: TryFromRequest + 'static, Req::Error: IntoResponse + 'static, Resp: IntoResponse + 'static,

Register a handler at the path for the specified HTTP method.

pub fn add_async<F, Fut, I, O>( &mut self, path: &str, method: Method, handler: F )
where F: Fn(I, Params) -> Fut + 'static, Fut: Future<Output = O> + 'static, I: TryFromRequest + 'static, I::Error: IntoResponse + 'static, O: IntoResponse + 'static,

Register an async handler at the path for the specified HTTP method.

pub fn get<F, Req, Resp>(&mut self, path: &str, handler: F)
where F: Fn(Req, Params) -> Resp + 'static, Req: TryFromRequest + 'static, Req::Error: IntoResponse + 'static, Resp: IntoResponse + 'static,

Register a handler at the path for the HTTP GET method.

pub fn get_async<F, Fut, Req, Resp>(&mut self, path: &str, handler: F)
where F: Fn(Req, Params) -> Fut + 'static, Fut: Future<Output = Resp> + 'static, Req: TryFromRequest + 'static, Req::Error: IntoResponse + 'static, Resp: IntoResponse + 'static,

Register an async handler at the path for the HTTP GET method.

pub fn head<F, Req, Resp>(&mut self, path: &str, handler: F)
where F: Fn(Req, Params) -> Resp + 'static, Req: TryFromRequest + 'static, Req::Error: IntoResponse + 'static, Resp: IntoResponse + 'static,

Register a handler at the path for the HTTP HEAD method.

pub fn head_async<F, Fut, Req, Resp>(&mut self, path: &str, handler: F)
where F: Fn(Req, Params) -> Fut + 'static, Fut: Future<Output = Resp> + 'static, Req: TryFromRequest + 'static, Req::Error: IntoResponse + 'static, Resp: IntoResponse + 'static,

Register an async handler at the path for the HTTP HEAD method.

pub fn post<F, Req, Resp>(&mut self, path: &str, handler: F)
where F: Fn(Req, Params) -> Resp + 'static, Req: TryFromRequest + 'static, Req::Error: IntoResponse + 'static, Resp: IntoResponse + 'static,

Register a handler at the path for the HTTP POST method.

pub fn post_async<F, Fut, Req, Resp>(&mut self, path: &str, handler: F)
where F: Fn(Req, Params) -> Fut + 'static, Fut: Future<Output = Resp> + 'static, Req: TryFromRequest + 'static, Req::Error: IntoResponse + 'static, Resp: IntoResponse + 'static,

Register an async handler at the path for the HTTP POST method.

pub fn delete<F, Req, Resp>(&mut self, path: &str, handler: F)
where F: Fn(Req, Params) -> Resp + 'static, Req: TryFromRequest + 'static, Req::Error: IntoResponse + 'static, Resp: IntoResponse + 'static,

Register a handler at the path for the HTTP DELETE method.

pub fn delete_async<F, Fut, Req, Resp>(&mut self, path: &str, handler: F)
where F: Fn(Req, Params) -> Fut + 'static, Fut: Future<Output = Resp> + 'static, Req: TryFromRequest + 'static, Req::Error: IntoResponse + 'static, Resp: IntoResponse + 'static,

Register an async handler at the path for the HTTP DELETE method.

pub fn put<F, Req, Resp>(&mut self, path: &str, handler: F)
where F: Fn(Req, Params) -> Resp + 'static, Req: TryFromRequest + 'static, Req::Error: IntoResponse + 'static, Resp: IntoResponse + 'static,

Register a handler at the path for the HTTP PUT method.

pub fn put_async<F, Fut, Req, Resp>(&mut self, path: &str, handler: F)
where F: Fn(Req, Params) -> Fut + 'static, Fut: Future<Output = Resp> + 'static, Req: TryFromRequest + 'static, Req::Error: IntoResponse + 'static, Resp: IntoResponse + 'static,

Register an async handler at the path for the HTTP PUT method.

pub fn patch<F, Req, Resp>(&mut self, path: &str, handler: F)
where F: Fn(Req, Params) -> Resp + 'static, Req: TryFromRequest + 'static, Req::Error: IntoResponse + 'static, Resp: IntoResponse + 'static,

Register a handler at the path for the HTTP PATCH method.

pub fn patch_async<F, Fut, Req, Resp>(&mut self, path: &str, handler: F)
where F: Fn(Req, Params) -> Fut + 'static, Fut: Future<Output = Resp> + 'static, Req: TryFromRequest + 'static, Req::Error: IntoResponse + 'static, Resp: IntoResponse + 'static,

Register an async handler at the path for the HTTP PATCH method.

pub fn options<F, Req, Resp>(&mut self, path: &str, handler: F)
where F: Fn(Req, Params) -> Resp + 'static, Req: TryFromRequest + 'static, Req::Error: IntoResponse + 'static, Resp: IntoResponse + 'static,

Register a handler at the path for the HTTP OPTIONS method.

pub fn options_async<F, Fut, Req, Resp>(&mut self, path: &str, handler: F)
where F: Fn(Req, Params) -> Fut + 'static, Fut: Future<Output = Resp> + 'static, Req: TryFromRequest + 'static, Req::Error: IntoResponse + 'static, Resp: IntoResponse + 'static,

Register an async handler at the path for the HTTP OPTIONS method.

pub fn new() -> Self

Construct a new Router.