Crate spin_sdk

The Rust Spin SDK.

Modules

config

Implementation of the spin config interface.

http

Helpers for building Spin HTTP components. These are convenience helpers, and the types in this module are based on the http crate.

key_value

Key/Value storage. Spin key-value persistent storage

llm

Large Language Model APIs

mysql

Implementation of the Spin MySQL database interface. Conversions between Rust, WIT and MySQL types.

outbound_http

Outbound HTTP request functionality.

pg

Implementation of the spin postgres db interface. Conversions between Rust, WIT and Postgres types.

redis

Implementation of the spin redis interface.

sqlite

SQLite storage.

Macros

export_reactor

Declares the export of the component’s world for the given type.

http_router

A macro to help with constructing a Router from a stream of tokens.

Attribute Macros

http_component

Exports the procedural macros for writing handlers for Spin components. The entrypoint to a Spin HTTP component written in Rust.

redis_component

Exports the procedural macros for writing handlers for Spin components. Generates the entrypoint to a Spin Redis component written in Rust.