Crate spin_sdk

The Rust Spin SDK.

Modules

config

Implementation of the spin config interface.

http

Helpers for building Spin HTTP components. These are convenience helpers, and the types in this module are based on the http crate.

mysql

Implementation of the Spin MySQL database interface. Conversions between Rust, WIT and MySQL types.

outbound_http

Outbound HTTP request functionality.

pg

Implementation of the spin postgres db interface. Conversions between Rust, WIT and Postgres types.

redis

Implementation of the spin redis interface.

Attribute Macros

http_component

The entrypoint to a Spin HTTP component written in Rust.

redis_component

Generates the entrypoint to a Spin Redis component written in Rust.