Crate spin_sdk

The Rust Spin SDK.

Modules

http

Helpers for building Spin wasi-http components.

key_value

Key/Value storage. Spin key-value persistent storage

llm

Large Language Model APIs

mysql

Implementation of the Spin MySQL database interface. Conversions between Rust, WIT and MySQL types.

pg

Implementation of the spin postgres db interface. Conversions between Rust, WIT and Postgres types.

redis

Implementation of the spin redis interface.

sqlite

SQLite storage.

variables

Macros

http_router

A macro to help with constructing a Router from a stream of tokens.

Attribute Macros

http_component

Exports the procedural macros for writing handlers for Spin components. The entrypoint to a WASI HTTP component written in Rust.

redis_component

Exports the procedural macros for writing handlers for Spin components. Generates the entrypoint to a Spin Redis component written in Rust.