Module spin_sdk::http::conversions

Traits for converting between the various types

Enums

IncomingRequestError

An error converting an IncomingRequest

Traits

FromBody

A trait from converting from a body

IntoBody

A trait for any type that can be turned into a Response body

IntoHeaders

A trait for any type that can be turned into Response headers

IntoResponse

A trait for any type that can be turned into a Response

IntoStatusCode

A trait for any type that can be turned into a Response status code

TryFromBody

A trait for converting from a body or failing

TryFromIncomingRequest

A trait for trying to convert from an IncomingRequest to the implementing type

TryFromIncomingResponse

A trait for converting from an IncomingRequest

TryFromRequest

A trait for any type that can be constructor from a Request

TryIntoBody

A trait for any type that can be turned into a Response body or fail

TryIntoOutgoingRequest

A trait for converting a type into an OutgoingRequest

TryIntoRequest

Turn a type into a Request

TryNonRequestFromRequest

A hack that allows us to do blanket impls for T where T: TryFromRequest for all types T except for Request.